Feature: Add CRUD functionality for albums and seed initial artist and album data in data.sql

+++ src/main/java/com/mtovar/musicat/service/RecordingService.java
@@ -1,6 +1,5 @@
 package com.mtovar.musicat.service;
 
-import com.mtovar.musicat.model.entity.Artist;
 import com.mtovar.musicat.model.entity.Recording;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
@@ -37,8 +36,8 @@ public class RecordingService {
         if (recording.getTitle() == null || recording.getTitle().isEmpty()) {
             throw new IllegalArgumentException("Recording name cannot be null or empty");
         }
-        if (recording.getDate() == null) {
-            throw new IllegalArgumentException("Recording date cannot be null");
+        if (recording.getYear() < 1857 || recording.getYear() > 2025) {
+            throw new IllegalArgumentException("Recording year must be between 1857 and the current year");
         }
         if (recordingRepository.existsByTitle(recording.getTitle())) {
             throw new IllegalArgumentException("Recording with name '" + recording.getTitle() + "' already exists");
@@ -64,21 +63,22 @@ public class RecordingService {
         if (recording.getTitle() == null || recording.getTitle().isEmpty()) {
             throw new IllegalArgumentException("Recording title cannot be null or empty");
         }
-        if (recording.getDate() == null) {
-            throw new IllegalArgumentException("Recording date cannot be null");
+        if (recording.getYear() < 1857 || recording.getYear() > 2025) {
+            throw new IllegalArgumentException("Recording year must be between 1857 and the current year");
         }
         if (!existingRecording.getTitle().equals(recording.getTitle()) && recordingRepository.existsByTitle(recording.getTitle())) {
             throw new IllegalArgumentException("Recording with name '" + recording.getTitle() + "' already exists");
         }
 
         existingRecording.setTitle(recording.getTitle());
-        existingRecording.setDate(recording.getDate());
+        existingRecording.setYear(recording.getYear());
         return recordingRepository.save(recording);
     }
 
     public void delete(Long id) {
-        Recording recording = recordingRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Recording not found with id: " + id));
+        if (!recordingRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recording not found with id: " + id);
+        }
         recordingRepository.deleteById(id);
     }
 }
